+++ CHANGELOG.md
@@ -8,6 +8,10 @@ the following form: YYYY.0M.0D.
 
 ## 2024.11.14
 
+### Added
+
+- Designed kernel implementation for HealthSystem component.
+
 ### Updated
 
 - Added `maxHealth` and `updateMaxHealth` Kernel methods